requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
bleach==5.0.1
bson==0.5.10
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
discord.py==2.0.1
docutils==0.19
frozenlist==1.3.1
halo==0.0.31
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
jaraco.classes==3.2.2
jsonc-parser==1.1.5
keyring==23.9.3
log-symbols==0.0.14
more-itertools==8.14.0
multidict==6.0.2
packaging==21.3
Pillow==9.2.0
pkginfo==1.8.3
pluggy==1.0.0
py==1.11.0
pyfiglet==0.8.post1
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-runner==6.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
readme-renderer==37.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
spinners==0.0.24
termcolor==2.0.1
tomli==2.0.1
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
yarl==1.8.1
zipp==3.8.1
